pyproject.toml
[project]
name = "bank_widjet"
version = "0.1.0"
description = ""
authors = [
    {name = "Rakhimzhanov Aman",email = "amanrahim2@gmail.com"}
]

requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pandas (>=2.3.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
black = "^25.1.0"
mypy = "^1.17.0"
flake8 = "^7.3.0"
isort = "^6.0.1"



[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"



[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(\.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
ignore_missing_imports = true
exclude = 'venv'





